requirements.txt
requests==2.22.0
pytz==2019.1
pandas==0.24.2
lxml==4.2.5
beautifulsoup4==4.9.0
python_dateutil==2.8.1
telethon==1.13.0
timezonefinder==4.4.0
PyYAML==5.3.1
